refactor: move add company component out of goal directory
